set user book relationship

+++ book_api/app/Http/Controllers/BookController.php
@@ -4,32 +4,51 @@ namespace App\Http\Controllers;
 
 use App\Models\Book;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class BookController extends Controller
 {
     public function index()
     {
-        return Book::with(['author', 'category'])->get();
+        return Book::with('user')->latest()->get();
+
     }
     public function store(Request $request)
     {
-        $book = Book::create($request->all());
-        return $book;
+        $book = $request->user()->books()->create($request->all());
+        return response()->json([
+            'message' => 'Book başarıyla kaydedildi',
+            'book' => $book,
+            'user' => $book->user,
+        ]);
+
     }
     public function show(Book $book)
     {
-        return $book;
+        return response()->json([
+            'book' => $book,
+            'user' => $book->user,
+        ]);
     }
     public function update(Request $request, Book $book)
     {
+        Gate::authorize('modify', $book);
+
         $book->update($request->all());
-        return $book;
+        return response()->json([
+            'book' => $book,
+            'message' => 'Book güncellendi',
+        ]);
+
     }
 
     public function destroy(Book $book)
     {
+        Gate::authorize('modify', $book);
 
         $book->delete();
-        return ["message" => "Book deleted successfully"];
+        return response()->json([
+            'message' => 'Book Silindi',
+        ]);
     }
 }
